Split `Relation` and `processRelations`
